requirements.txt
numpy
Pillow
nifpga
PySide2
qimage2ndarray
tifffile
imageio
pytest
pytest-qt
pyyaml